Rename admonition extension module
